# back/src/routes/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Gerenciamento de Tarefas",
    "version": "1.0.0",
    "description": "API para gerenciar uma lista de tarefas."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Lista todas as tarefas",
        "responses": {
          "200": {
            "description": "Lista de tarefas retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria uma nova tarefa",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTask"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarefa criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação"
          }
        }
      }
    },
    "/tasks/{id}": {
      "put": {
        "summary": "Atualiza o status de uma tarefa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da tarefa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarefa atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada"
          }
        }
      },
      "delete": {
        "summary": "Exclui uma tarefa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da tarefa"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa excluída com sucesso"
          },
          "404": {
            "description": "Tarefa não encontrada"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
      "UpdateTask": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["concluída", "não concluída"]
          }
        }
      }
    }
  }
}